Move Constraint function annotation to SPI
