+++ core/trino-main/src/main/java/io/trino/operator/scalar/ColorFunctions.java
@@ -16,12 +16,12 @@ package io.trino.operator.scalar;
 import com.google.common.annotations.VisibleForTesting;
 import io.airlift.slice.Slice;
 import io.trino.spi.TrinoException;
+import io.trino.spi.function.Constraint;
 import io.trino.spi.function.LiteralParameters;
 import io.trino.spi.function.ScalarFunction;
 import io.trino.spi.function.SqlType;
 import io.trino.spi.type.StandardTypes;
 import io.trino.type.ColorType;
-import io.trino.type.Constraint;
 
 import java.awt.Color;
