+++ core/trino-main/src/main/java/io/trino/operator/scalar/MathFunctions.java
@@ -23,6 +23,7 @@ import io.trino.metadata.SqlScalarFunction;
 import io.trino.spi.TrinoException;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.SqlMap;
+import io.trino.spi.function.Constraint;
 import io.trino.spi.function.Convention;
 import io.trino.spi.function.Description;
 import io.trino.spi.function.LiteralParameter;
@@ -36,7 +37,6 @@ import io.trino.spi.type.Int128;
 import io.trino.spi.type.StandardTypes;
 import io.trino.type.BlockTypeOperators.BlockPositionHashCode;
 import io.trino.type.BlockTypeOperators.BlockPositionIsIdentical;
-import io.trino.type.Constraint;
 import org.apache.commons.math3.distribution.BetaDistribution;
 import org.apache.commons.math3.distribution.TDistribution;
 import org.apache.commons.math3.special.Erf;
