+++ core/trino-main/src/main/java/io/trino/operator/scalar/Re2JRegexpFunctions.java
@@ -19,13 +19,13 @@ import io.airlift.slice.Slices;
 import io.trino.re2j.Matcher;
 import io.trino.spi.TrinoException;
 import io.trino.spi.block.Block;
+import io.trino.spi.function.Constraint;
 import io.trino.spi.function.Description;
 import io.trino.spi.function.LiteralParameters;
 import io.trino.spi.function.ScalarFunction;
 import io.trino.spi.function.SqlNullable;
 import io.trino.spi.function.SqlType;
 import io.trino.spi.type.StandardTypes;
-import io.trino.type.Constraint;
 import io.trino.type.Re2JRegexp;
 import io.trino.type.Re2JRegexpType;
