+++ core/trino-main/src/main/java/io/trino/operator/scalar/UrlFunctions.java
@@ -19,13 +19,13 @@ import com.google.common.net.UrlEscapers;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
 import io.trino.spi.TrinoException;
+import io.trino.spi.function.Constraint;
 import io.trino.spi.function.Description;
 import io.trino.spi.function.LiteralParameters;
 import io.trino.spi.function.ScalarFunction;
 import io.trino.spi.function.SqlNullable;
 import io.trino.spi.function.SqlType;
 import io.trino.spi.type.StandardTypes;
-import io.trino.type.Constraint;
 import jakarta.annotation.Nullable;
 
 import java.io.UnsupportedEncodingException;
