+++ core/trino-main/src/main/java/io/trino/operator/scalar/timestamp/TimestampOperators.java
@@ -13,12 +13,12 @@
  */
 package io.trino.operator.scalar.timestamp;
 
+import io.trino.spi.function.Constraint;
 import io.trino.spi.function.LiteralParameters;
 import io.trino.spi.function.ScalarOperator;
 import io.trino.spi.function.SqlType;
 import io.trino.spi.type.LongTimestamp;
 import io.trino.spi.type.StandardTypes;
-import io.trino.type.Constraint;
 import org.joda.time.DateTimeField;
 import org.joda.time.chrono.ISOChronology;
