+++ core/trino-main/src/main/java/io/trino/operator/scalar/timestamptz/ToIso8601.java
@@ -15,12 +15,12 @@ package io.trino.operator.scalar.timestamptz;
 
 import io.airlift.slice.Slice;
 import io.trino.spi.TrinoException;
+import io.trino.spi.function.Constraint;
 import io.trino.spi.function.LiteralParameter;
 import io.trino.spi.function.LiteralParameters;
 import io.trino.spi.function.ScalarFunction;
 import io.trino.spi.function.SqlType;
 import io.trino.spi.type.LongTimestampWithTimeZone;
-import io.trino.type.Constraint;
 import io.trino.type.DateTimes;
 
 import java.time.Instant;
