+++ core/trino-spi/src/main/java/io/trino/spi/function/Constraint.java
@@ -11,16 +11,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.type;
+package io.trino.spi.function;
 
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Constraints
+@Repeatable(Constraints.class)
+public @interface Constraint
 {
-    Constraint[] value();
+    String variable() default "";
+
+    String expression() default "";
 }
